switch web scrapes
