+++ src/autolabel/labeler.py
@@ -272,10 +272,8 @@ class LabelingAgent:
                 max_input_tokens=self.llm.DEFAULT_CONTEXT_LENGTH,
                 get_num_tokens=self.llm.get_num_tokens,
             )
-            logger.info(f"Prompt: {final_prompt}")
 
             response = await self.llm.label([final_prompt])
-            logger.info(f"Response: {response}")
             for i, generations, error, latency in zip(
                 range(len(response.generations)),
                 response.generations,
