+++ src/autolabel/transforms/webpage_scrape.py
@@ -1,4 +1,3 @@
-from urllib.parse import urlparse
 from autolabel.transforms.schema import (
     TransformType,
     TransformError,
@@ -10,21 +9,25 @@ import asyncio
 import logging
 import pandas as pd
 import ssl
-from langchain_community.document_transformers import Html2TextTransformer
-from langchain.docstore.document import Document
+
 from autolabel.cache import BaseCache
-from scrapingbee import ScrapingBeeClient
 
 logger = logging.getLogger(__name__)
 
 MAX_RETRIES = 5
+MAX_KEEPALIVE_CONNECTIONS = 20
+MAX_CONNECTIONS = 100
 BACKOFF = 2
-DEFAULT_TIMEOUT = 60000  # in milliseconds
+HEADERS = {}
+HTML_PARSER = "html.parser"
 
 
 class WebpageScrape(BaseTransform):
     COLUMN_NAMES = [
         "content_column",
+        "content_in_bytes_column",
+        "soup_column",
+        "metadata_column",
     ]
 
     def __init__(
@@ -32,24 +35,59 @@ class WebpageScrape(BaseTransform):
         cache: BaseCache,
         output_columns: Dict[str, Any],
         url_column: str,
-        timeout: int = DEFAULT_TIMEOUT,
-        scrapingbee_api_key: str = None,
+        timeout: int = 60,
+        headers: Dict[str, str] = HEADERS,
     ) -> None:
         super().__init__(cache, output_columns)
         self.url_column = url_column
+        self.headers = headers
         self.max_retries = MAX_RETRIES
-        self.timeout = DEFAULT_TIMEOUT
-        self.html2text_transformer = Html2TextTransformer()
-        self.api_key = scrapingbee_api_key
-        self.client = ScrapingBeeClient(api_key=self.api_key)
-        self.scrapingbee_params = {
-            "timeout": self.timeout,
-        }
+        try:
+            from bs4 import BeautifulSoup
+            import httpx
+
+            if not headers.get("User-Agent"):
+                from fake_useragent import UserAgent
+
+                headers["User-Agent"] = UserAgent().random
+
+            self.httpx = httpx
+            self.timeout_time = timeout
+            self.timeout = httpx.Timeout(timeout)
+            limits = httpx.Limits(
+                max_keepalive_connections=MAX_KEEPALIVE_CONNECTIONS,
+                max_connections=MAX_CONNECTIONS,
+                keepalive_expiry=timeout,
+            )
+            self.client = httpx.AsyncClient(
+                timeout=self.timeout, limits=limits, follow_redirects=True
+            )
+            self.client_with_no_verify = httpx.AsyncClient(
+                timeout=self.timeout, limits=limits, follow_redirects=True, verify=False
+            )
+            self.beautiful_soup = BeautifulSoup
+        except ImportError:
+            raise ImportError(
+                "BeautifulSoup, httpx and fake_useragent are required to use the webpage transform. Please install them with the following command: pip install beautifulsoup4 httpx fake_useragent"
+            )
 
     def name(self) -> str:
         return TransformType.WEBPAGE_SCRAPE
 
-    async def _load_url(self, url: str, retry_count=0) -> str:
+    def _load_metadata(self, url, soup) -> Dict[str, Any]:
+        metadata = {"url": url}
+        if soup.find("title"):
+            metadata["title"] = soup.find("title").get_text()
+        for meta in soup.find_all("meta"):
+            if meta.get("name") and meta.get("content"):
+                metadata[meta.get("name")] = meta.get("content")
+            elif meta.get("property") and meta.get("content"):
+                metadata[meta.get("property")] = meta.get("content")
+        return metadata
+
+    async def _load_url(
+        self, url: str, verify=True, headers=HEADERS, retry_count=0
+    ) -> Dict[str, Any]:
         if retry_count >= self.max_retries:
             logger.warning(f"Max retries reached for URL: {url}")
             raise TransformError(
@@ -57,38 +95,63 @@ class WebpageScrape(BaseTransform):
             )
 
         try:
-            response = self.client.get(url, params=self.scrapingbee_params)
-            documents = [
-                Document(page_content=response.content, metadata={"source": url})
-            ]
-            text = self.html2text_transformer.transform_documents(documents)[
-                0
-            ].page_content
-            return text
-        except Exception as e:
+            client = self.client
+            if not verify:
+                client = self.client_with_no_verify
+            response = await client.get(url, headers=headers)
+
+            # TODO: Add support for other parsers
+            content_bytes = response.content
+            soup = self.beautiful_soup(content_bytes, HTML_PARSER)
+            return {
+                "content": soup.get_text(),
+                "content_bytes": content_bytes,
+                "soup": soup,
+                "metadata": self._load_metadata(url, soup),
+            }
+        except self.httpx.ConnectTimeout as e:
+            logger.error(f"Timeout when fetching content from URL: {url}")
+            raise TransformError(
+                TransformErrorType.TRANSFORM_TIMEOUT,
+                "Timeout when fetching content from URL",
+            )
+        except ssl.SSLCertVerificationError as e:
+            logger.warning(
+                f"SSL verification error when fetching content from URL: {url}, retrying with verify=False"
+            )
             await asyncio.sleep(BACKOFF**retry_count)
-            return await self._load_url(url, retry_count=retry_count + 1)
+            return await self._load_url(
+                url, verify=False, headers=headers, retry_count=retry_count + 1
+            )
+        except Exception as e:
+            logger.error(f"Error fetching content from URL: {url}. Exception: {e}")
+            raise e
 
     async def _apply(self, row: Dict[str, Any]) -> Dict[str, Any]:
         url = row[self.url_column]
+        url_response_data = {}
         if pd.isna(url) or url == self.NULL_TRANSFORM_TOKEN:
             raise TransformError(
                 TransformErrorType.INVALID_INPUT,
                 f"Empty url in row {row}",
             )
-        if not urlparse(url).scheme:
-            url = f"https://{url}"
-        url_response_text = await self._load_url(url)
+        else:
+            url_response_data = await self._load_url(url)
 
         transformed_row = {
-            self.output_columns["content_column"]: url_response_text,
+            self.output_columns["content_column"]: url_response_data.get("content"),
+            self.output_columns["content_in_bytes_column"]: url_response_data.get(
+                "content_bytes"
+            ),
+            self.output_columns["soup_column"]: url_response_data.get("soup"),
+            self.output_columns["metadata_column"]: url_response_data.get("metadata"),
         }
+
         return self._return_output_row(transformed_row)
 
     def params(self):
         return {
             "url_column": self.url_column,
             "output_columns": self.output_columns,
-            "timeout": self.timeout,
-            "scrapingbee_api_key": self.api_key,
+            "timeout": self.timeout_time,
         }
